+++ docs/README.md
@@ -8,6 +8,6 @@
 - ✅ 게임 종료 출력 구현
 - ✅ 컴퓨터의 랜덤수 생성
 - ✅ 사용자 입력
-- 사용자가 잘못된 값을 입력한 경우 예외 처리
-- 재시작 버튼 구현
+- ✅ 사용자가 잘못된 값을 입력한 경우 예외 처리
+- ✅ 재시작 버튼 구현
